requirements.txt
alembic==1.4.3
aniso8601==8.0.0
appdirs==1.4.4
apturl==0.5.2
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
neo4j==4.1.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pbr==5.5.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-editor==1.0.4
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
shortuuid==1.0.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.20
stevedore==3.2.2
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.1.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
Werkzeug==0.16.1
xkit==0.0.0

py2neo~=2021.0.1